feat: added factorial of decimal exception

+++ Math_Func.py
@@ -1,4 +1,4 @@
-def fac(n: float):
+def fac(n: int):
     fac = 1
     for i in range(1,n+1):
         fac *= i
